requirements.txt
Django==2.2.5
mysqlclient==1.4.4
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
